requirements.txt
discord>=2.3.2
rapidfuzz>=3.9.3
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
markdownify>=0.12.1
python-dotenv>=1.0.1